src/analise_sentimento_hb20_onix.py
import pandas as pd
from sklearn.model_selection import train_test_split
from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.naive_bayes import MultinomialNB
from sklearn.metrics import classification_report, confusion_matrix
import numpy as np

# Nome do arquivo CSV ROTULADO da etapa anterior
NOME_ARQUIVO_ROTULADO = 'tweets_hb20_onix_2000_rotulado.csv'

# Carregar o DataFrame rotulado
try:
    df = pd.read_csv(NOME_ARQUIVO_ROTULADO, encoding='utf-8')
    TEXT_COLUMN = 'content_processado' # Coluna limpa e sem stopwords
    LABEL_COLUMN_BRUTA = 'sentiment_label'
    
    # Excluir linhas onde o texto processado está vazio
    df.dropna(subset=[TEXT_COLUMN, LABEL_COLUMN_BRUTA], inplace=True)
    
    print(f"✅ Arquivo rotulado '{NOME_ARQUIVO_ROTULADO}' carregado. Linhas válidas: {len(df)}")
    print("--- Iniciando a Modelagem (3 Classes) ---")
    
except FileNotFoundError:
    print(f"❌ ERRO: Arquivo '{NOME_ARQUIVO_ROTULADO}' não encontrado. Verifique a execução anterior.")
    exit()

# ----------------------------------------------------
# 1. PRÉ-MODELAGEM: SIMPLIFICAÇÃO PARA 3 CLASSES
# ----------------------------------------------------

# Mapeamento para simplificar as 4 classes heurísticas para 3 classes padrão do TCC
mapeamento_sentimento = {
    'Positivo': 'Positivo',
    'Negativo': 'Negativo',
    # Agrupando ruído, irrelevância e ambiguidade em uma única classe Neutra
    'Neutro/Ruído': 'Neutro',
    'Neutro/Conflito': 'Neutro'
}

df['sentiment_final'] = df[LABEL_COLUMN_BRUTA].map(mapeamento_sentimento)

# Definir as colunas para a modelagem
X = df[TEXT_COLUMN]
y = df['sentiment_final']

print("\n--- Distribuição Final das 3 Classes ---")
print(y.value_counts(normalize=True).mul(100).round(1).astype(str) + '%')

# ----------------------------------------------------
# 2. DIVISÃO DOS DADOS (TREINO E TESTE)
# ----------------------------------------------------

X_train, X_test, y_train, y_test = train_test_split(
    X, y, 
    test_size=0.2, 
    random_state=42, 
    stratify=y 
)

# ----------------------------------------------------
# 3. VETORIZAÇÃO (TF-IDF)
# ----------------------------------------------------

print("\n--- Vetorização dos Dados (TF-IDF) ---")
tfidf = TfidfVectorizer(max_features=5000) 

# Ajustar e transformar (fit_transform) no treino
X_train_vectorized = tfidf.fit_transform(X_train).toarray()

# Transformar (transform) no teste
X_test_vectorized = tfidf.transform(X_test).toarray()

# ----------------------------------------------------
# 4. TREINAMENTO DO CLASSIFICADOR (NAIVE BAYES)
# ----------------------------------------------------

print("--- Treinando o Modelo Naive Bayes (MNB) ---")
classifier = MultinomialNB()
classifier.fit(X_train_vectorized, y_train)

print("✅ Treinamento concluído!")

# ----------------------------------------------------
# 5. AVALIAÇÃO DO DESEMPENHO
# ----------------------------------------------------

y_pred = classifier.predict(X_test_vectorized)

print("\n=======================================================")
print("  RESULTADOS DA AVALIAÇÃO DO MODELO (3 CLASSES)  ")
print("=======================================================")

# A) MATRIZ DE CONFUSÃO
print("\n[A] Matriz de Confusão:")
cm = confusion_matrix(y_test, y_pred, labels=classifier.classes_)
cm_df = pd.DataFrame(cm, index=classifier.classes_, columns=classifier.classes_)
print(cm_df)

# B) RELATÓRIO DE CLASSIFICAÇÃO
print("\n[B] Relatório de Classificação (Precision, Recall, F1-Score):")
print(classification_report(y_test, y_pred))

# C) PRECISÃO GLOBAL
accuracy = np.mean(y_pred == y_test)
print(f"\n[C] Precisão Global do Modelo (Accuracy): {accuracy:.4f}")
print("=======================================================")

# ----------------------------------------------------
# 6. DEMONSTRAÇÃO PRÁTICA
# ----------------------------------------------------

print("\n--- Demonstração Prática (Teste de Novas Frases) ---")
frases_teste_real = [
    "A suspensão dura desse carro é um ranço, que lixo, não aguento mais!", # Negativo
    "O design esportivo do HB20 é sensacional, valeu cada centavo. Topzera!", # Positivo
    "Meu foco agora é a liquidez na revenda, tô pensando no onix ou no kwid.", # Neutro
    "Vi um onix vermelho e lembrei que tenho que comprar pão, ué.", # Neutro
    "Adorei o câmbio travando, é super de boa. Sarcasmo total! 😭" # Negativo (Ironia)
]

# (Opcional: Limpar e vetorizar novas frases para o teste)
# O código real aqui precisaria limpar as novas frases, mas manteremos o foco
# na vetorização para fins de demonstração rápida.

X_new_vectorized = tfidf.transform(frases_teste_real)
novas_predicoes = classifier.predict(X_new_vectorized)

print("\nResultados das Novas Predições:")
for frase, predicao in zip(frases_teste_real, novas_predicoes):
    print(f"Frase: '{frase[:50]}...' -> Predição: {predicao}")

print("\n✅ O projeto de Análise de Sentimentos para seu TCC está completo e otimizado para apresentação.")