requirements.txt
# Manipulação de dados
pandas==2.2.2
numpy==1.26.4

# Coleta de tweets
snscrape==0.7.0

# Pré-processamento e NLP
scikit-learn==1.5.0
nltk==3.8.1
regex==2024.5.15

# Modelos pré-treinados
transformers==4.41.2
torch==2.3.0
sentencepiece==0.2.0

# Visualização
matplotlib==3.9.0
seaborn==0.13.2

# Dashboard
streamlit==1.36.0

# Banco de dados MySQL
mysql-connector-python==8.0.33
sqlalchemy==2.0.30
